backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic[email] # Changed from pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart